Create README.md
Project Description
This project was created as a practice exercise to enhance HTML and CSS skills by following a design provided in a Figma file. It serves as an opportunity to apply web development concepts and techniques in a real-world scenario.

The main objective of this project was to meticulously recreate the design and layout provided in the Figma file. By doing so, the project aimed to strengthen proficiency in HTML structure, and CSS styling.

Key Features
Design Implementation: The project faithfully replicates the design elements, including typography, colors, spacing, and overall visual hierarchy specified in the original Figma file.
HTML Structure: Careful attention was given to organizing the HTML structure to reflect the semantic nature of the content. 
CSS Styling: The project showcases CSS styling techniques, such as selectors, classes, IDs, and cascading rules, to accurately reproduce the visual appearance of the Figma design.
Responsiveness: The layout was developed with responsiveness in mind, ensuring optimal display across different screen sizes and devices. Media queries and responsive design principles were employed to achieve a fluid and adaptive user experience.
Unique Aspects
Figma-to-Code Workflow: This project allowed for hands-on experience in converting a design file into a functional web page. The process involved analyzing the design, extracting visual details, and translating them into HTML and CSS code.
Attention to Detail: The project emphasized precision and attention to detail, aiming for pixel-perfect alignment and consistent styling throughout the webpage. This attention to detail helps enhance the overall user experience and showcases a commitment to quality.
Learning Opportunity: Through this project, valuable insights and knowledge were gained about the workflow of turning a design into a web page, improving HTML and CSS skills, and understanding the challenges and considerations involved in web development.